* did some cleaning of the source * did some speedup's in the spell class * did some -walling to remove unneeded stuff / old stuff * replaced a switch -> case for a if (last time I checked a single if was faster than a single switch) * removed some -wall warnings * did some typedef work, I hope this makes the source a bit cleaner Note: most of these things are NTY but should not give problems. Please report the problems you have to me.

+++ src/game/NPCHandler.cpp
@@ -185,7 +185,7 @@ void WorldSession::SendTrainerList(Creature* pCreature)
 					RequiredLevel = 10;
 					break;
 				default:
-					RequiredLevel = -1;
+					RequiredLevel = uint32(-1);
 				}
 			}
